+++ src/collective/folderishtypes/interfaces.py
@@ -1,6 +1,11 @@
 from zope.interface import Interface
 
 
+class IThemeSpecific(Interface):
+    """Marker interface that defines a Zope 3 browser layer.
+    """
+
+
 class IFolderishType(Interface):
     """Marker interface
     """