+++ src/collective/folderishtypes/upgrades.py
@@ -1,0 +1,8 @@
+
+
+def import_browserlayer(context):
+    context.runImportStepFromProfile(
+        'profile-collective.folderishtypes:default',
+        'browserlayer',
+    )
+